Move stuff around. Use pytest.
