+++ .gitignore
@@ -1,5 +1,8 @@
 .vscode
 
+.env
+venv
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
