+++ clean.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-# This script removes the virtual Python environment and 
+# This script removes the virtual Python environment and the environment file.
 
 read -p "This will remove the virtual Python environment and the database credentials. Are you sure? (y/n) " -n 1 -r
 echo
