--- test_ml/.gitignore
@@ -1,2 +1,0 @@
-.env
-venv
